requirements.txt
tweepy>=4.16.0
praw==7.7.1
google-api-python-client==2.108.0
pymongo==4.6.0
pyspark==3.5.0
tensorflow==2.20.0
pandas>=2.2.0
numpy>=1.26.4
scikit-learn>=1.5.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
nltk==3.8.1
textblob==0.17.1
schedule==1.2.0
psutil==5.9.8